% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_body.R
\name{dashboard_body}
\alias{dashboard_body}
\title{Shinydashboard Body}
\usage{
dashboard_body()
}
\description{
The body is where we will place our UI/modules. This function is a key component to the [shinydashboard::dashboardPage()]. In our dashboard we use an extension library to `shinydashboard` called `shindydashboardPlus` where we use [shinydashboardPlus::dashboardPagePlus()] for added features. In the [dashboard_page()] function you will find where these different shinydashboard components come together to form the dashboard.
}
\seealso{
Other Shiny Dashboard Functions: 
\code{\link{dashboard_header}()},
\code{\link{dashboard_page}()},
\code{\link{dashboard_sidebar}()}
}
\concept{Shiny Dashboard Functions}
