% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_header.R
\name{dashboard_header}
\alias{dashboard_header}
\title{Shinydashboard Header}
\usage{
dashboard_header()
}
\description{
The header is where we declare the title.
}
\seealso{
Other Shiny Dashboard Functions: 
\code{\link{dashboard_body}()},
\code{\link{dashboard_page}()},
\code{\link{dashboard_sidebar}()}
}
\concept{Shiny Dashboard Functions}
