% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_page.R
\name{dashboard_page}
\alias{dashboard_page}
\title{Shiny Dashboard Page}
\usage{
dashboard_page()
}
\description{
This function brings all the shinydashboard components together; [dashboard_header()], [dashboard_sidebar()], and [dashboard_body()].
}
\seealso{
Other Shiny Dashboard Functions: 
\code{\link{dashboard_body}()},
\code{\link{dashboard_header}()},
\code{\link{dashboard_sidebar}()}
}
\concept{Shiny Dashboard Functions}
