% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_sidebar.R
\name{dashboard_sidebar}
\alias{dashboard_sidebar}
\title{Shinydashboard Sidebar}
\usage{
dashboard_sidebar()
}
\description{
The sidebar is where we declare wanted menu items/tabs. This function is a key component to the [shinydashboard::dashboardPage()]. In our dashboard we use an extension library to `shinydashboard` called `shindydashboardPlus` where we use [shinydashboardPlus::dashboardPagePlus()] for added features. In the [dashboard_page()] function you will find where these different shinydashboard components come together to form the dashboard.
}
\seealso{
Other Shiny Dashboard Functions: 
\code{\link{dashboard_body}()},
\code{\link{dashboard_header}()},
\code{\link{dashboard_page}()}
}
\concept{Shiny Dashboard Functions}
