% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_bayes_deconv.R
\name{fct_bayes_deconv}
\alias{fct_bayes_deconv}
\title{Title}
\usage{
fct_bayes_deconv(
  .data,
  lo_cut = 0,
  hi_cut = 0.05,
  table = FALSE,
  updateProgress = NULL
)
}
\arguments{
\item{.data}{}
}
\description{
Title
}
