% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_model_select_bic.R
\name{fct_model_select_bic}
\alias{fct_model_select_bic}
\title{Title}
\usage{
fct_model_select_bic(data, grid, param = 2:5)
}
\arguments{
\item{data}{}

\item{grid}{}

\item{param}{}
}
\description{
Title
}
