% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_report_results.R
\name{fct_report_results}
\alias{fct_report_results}
\title{Title}
\usage{
fct_report_results(model, lo_cut = 0, hi_cut = 0.05, table = FALSE)
}
\arguments{
\item{model}{}

\item{lo_cut}{}

\item{hi_cut}{}

\item{table}{}
}
\description{
Title
}
