% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_background.R
\name{mod_background_ui}
\alias{mod_background_ui}
\title{background UI Function
#' @description A shiny Module.#'#' @param id,input,output,session Internal parameters for {shiny}.#'#' @noRd #'#' @importFrom shiny NS tagList}
\usage{
mod_background_ui(id)
}
\description{
background UI Function
#' @description A shiny Module.#'#' @param id,input,output,session Internal parameters for {shiny}.#'#' @noRd #'#' @importFrom shiny NS tagList
}
