% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_details.R
\name{mod_details_ui}
\alias{mod_details_ui}
\title{details UI Function
#' @description A shiny Module.#'#' @param id,input,output,session Internal parameters for {shiny}.
#' @noRd #'#' @importFrom shiny NS tagList}
\usage{
mod_details_ui(id)
}
\description{
details UI Function
#' @description A shiny Module.#'#' @param id,input,output,session Internal parameters for {shiny}.
#' @noRd #'#' @importFrom shiny NS tagList
}
