% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results_dt.R
\name{mod_results_dt_ui}
\alias{mod_results_dt_ui}
\title{results_dt UI Function#'#' @description A shiny Module.#'}
\usage{
mod_results_dt_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.#'#' @noRd 
#' @importFrom shiny NS tagList}
}
\description{
results_dt UI Function#'#' @description A shiny Module.#'
}
